==> ImmuneCellAnalysis/04a_Integrate_Immune.Tumor/CellChat/01_PrepForCellChat.Rmd <==
---
title: "PrepSeuratObjectsForCellChat"
output: html_document
---

```{r}
library(CellChat)
library(Seurat) 

options(stringsAsFactors = FALSE)

base_dir = "/Users/jlabelle/"
script_dir = "Dropbox (Partners HealthCare)/FilbinLab/data_analysis/Scripts/"
working_dir<- paste0(base_dir, 
                     "Dropbox (Partners HealthCare)/FilbinLab/",
                     "data_analysis/Projects/git_repos/Pedi_Immune/ImmuneCellAnalysis/04a_Integrate_Immune.Tumor/CellChat/")

analysis_dir<- paste0(working_dir, "/analysis/")
if(!dir.exists(analysis_dir)){dir.create(analysis_dir, recursive = TRUE)}
```

## Read in seurat objects: ped tcells, myeloid cells, tumor cells
```{r}
rank_use<-8
sharedBy<- "union"
cohort<- "pedOnly_nomuv63"

tcell_data_dir<- paste0("../../02a_Tcells/02_Ped.Adult/03_Visualize.Reannotate_programs/analysis/Reannotate_programs/",
                         "rank", rank_use, "_sharedBy", sharedBy, "/")
myeloid_data_dir<- paste0("../../01_Preprocessing/02_Counts_Preprocess/analysis/detailed_annotation/", cohort, "/")

tcell_seurat<- readRDS(paste0(tcell_data_dir, "ped_seurat_tcell_harmony.Rds"))
myeloid_seurat<- readRDS(paste0(myeloid_data_dir, "myeloid_seurat_noDC.B_harmony.Rds"))
load("../../../TumorCellAnnotation/AllSubtypesTogether/analysis/Merge_allSubtypes/AllTumorSeuratObj.Robj")
tumor_seurat<- tumor_merged; rm(tumor_merged)
```


## Merge 3 seurat objects together
```{r}
## Add common column for: 
## 1) high level cell type (tcell, myeloid, tumor)
tcell_seurat$highLevel_annot<- "Tcell"
myeloid_seurat$highLevel_annot<- "Myeloid"
tumor_seurat$highLevel_annot<- "Tumor"

## 2) detailed (myeloid, cd4/cd8 clonal/nonclonal, tumor cell type)
## For myeloid: may replace this with microglia/macrophage/monocyte distinction
tcell_seurat$detailed_annotForCellChat<- paste0(tcell_seurat$detailed_annot_byNMFrank8_mergedByunion, ".",
                                                tcell_seurat$clonal)
myeloid_seurat$detailed_annotForCellChat<-"Myeloid"
tumor_seurat$detailed_annotForCellChat<- gsub("cycling", "Cycling", tumor_seurat$CellAnnot)

## Add on any other colums needed before merging
tumor_seurat$sampleid<- tumor_seurat$sample

## Check that scaling is the same for all
head(colSums(myeloid_seurat@assays$RNA@counts))/head(colSums(tcell_seurat@assays$RNA@counts))
head(colSums(myeloid_seurat@assays$RNA@counts))/head(colSums(tumor_seurat@assays$RNA@counts))

## Match metadata to clean up merged seurat
commonColumns<- colnames(tumor_seurat@meta.data)[colnames(tumor_seurat@meta.data) %in% c(colnames(myeloid_seurat@meta.data),
                                                                                         colnames(tcell_seurat))]
tumor_seurat@meta.data<- tumor_seurat@meta.data[,commonColumns]
myeloid_seurat@meta.data<- myeloid_seurat@meta.data[,commonColumns]
tcell_seurat@meta.data<- tcell_seurat@meta.data[,commonColumns]

## Merge together
merged_seurat<- merge(tumor_seurat, c(myeloid_seurat, tcell_seurat))

## Run full seurat processing- not needed for cellChat, but nice for a high-level look
merged_seurat_meta<- merged_seurat@meta.data
merged_seurat<- RunFullSeurat(merged_seurat@assays$RNA@counts, RunHarmony = TRUE, samples=merged_seurat_meta$sampleid)
merged_seurat<- AddMetaData(merged_seurat, merged_seurat_meta[,!colnames(merged_seurat_meta) %in% colnames(merged_seurat@meta.data)])

saveRDS(merged_seurat, file=paste0(analysis_dir, "seurat_tumor.tcell.myeloid.Rds"))
```

## Add on myeloid program and tcell program
```{r}
all_seurat<- readRDS(paste0(analysis_dir, "seurat_tumor.tcell.myeloid.Rds"))

## Create named vector of all annotations
annots<- c(paste0(tcell_seurat$detailed_annot_byNMFrank8_mergedByunion, "_", tcell_seurat$NMF_rank8_annot_mergedByunion_reAnnot),
           as.character(myeloid_seurat$NMF_rank6_annotByAdult10X),
           tumor_seurat$CellAnnot)
names(annots)<- c(colnames(tcell_seurat), colnames(myeloid_seurat), colnames(tumor_seurat))

## Order vector of annotations the same way as in seurat object
sum(!names(annots) %in% colnames(all_seurat))
annots<- annots[colnames(all_seurat)]
sum(names(annots) != colnames(all_seurat))

## Add annotations onto seurat object
all_seurat$immunePrograms_tumorCellTypes<- unname(annots)

## Fix any annotations
all_seurat$immunePrograms_tumorCellTypes<- gsub("cycling", "Cycling", all_seurat$immunePrograms_tumorCellTypes)

saveRDS(all_seurat, file=paste0(analysis_dir, "seurat_tumor.tcell.myeloid.Rds"))
```

## Split into 1 seurat object per sample
```{r}
all_seurat<- readRDS(paste0(analysis_dir, "seurat_tumor.tcell.myeloid.Rds"))

## Number of cells in each cell type per sample, save
## This includes detailed annot (CD4, CD8, etc) as well as Tcell/myeloid programs
meta<- all_seurat@meta.data

  ## detailed annot
  nCells_bySample<- as.data.frame(table(meta$detailed_annotForCellChat, meta$sample))
  colnames(nCells_bySample)<-c("CellType", "Sample", "NCells")
  
  ## By program
  nProgramCells_bySample<- as.data.frame(table(meta$immunePrograms_tumorCellTypes, meta$sample))
  colnames(nProgramCells_bySample)<-c("CellType", "Sample", "NCells")

nCells_bySample<- rbind(nCells_bySample, nProgramCells_bySample[!nProgramCells_bySample$CellType %in% nCells_bySample$CellType,])  
saveRDS(nCells_bySample, file=paste0(analysis_dir, "nCellTypes_bySample.Rds"))

## Cycle through each sample, split seurat object
all_seurat_list<- lapply(unique(meta$sample), function(x){
  seurat_tmp<- subset(all_seurat, sample==x)
  print(paste0(x, " has ", ncol(seurat_tmp)," cells"))
  return(seurat_tmp)
})
names(all_seurat_list)<- unique(meta$sample)

saveRDS(all_seurat_list, file=paste0(analysis_dir, "seurat_list_bySample.Rds"))
```


## Write out comparison file- sender/receiver text file
file_with_comparisons should be a text file with 3 columns that follow the structure: comparison_name sender receiver
Can run all simultaneously, but takes too long to run- over 3 days
Split into multiple comparison files, just so they can run simultaneously

## By detailed cell type (including clonality): tumor cell types, myeloid, cd4 (clonal/nonclonal), cd8 (clonal/nonclonal)
```{r}
all_seurat<- readRDS(paste0(analysis_dir, "seurat_tumor.tcell.myeloid.Rds"))

## Get list of all cell types by high level cell type
tcell_types<- unique(all_seurat$detailed_annotForCellChat[all_seurat$highLevel_annot=="Tcell"])
tcell_types<- tcell_types[!grepl("Unclear", tcell_types)]
myeloid_types<- unique(all_seurat$detailed_annotForCellChat[all_seurat$highLevel_annot=="Myeloid"])
tumor_types<- unique(all_seurat$detailed_annotForCellChat[all_seurat$highLevel_annot=="Tumor"])

## Between high level cell type lists, get all pairwise combinations
tcell_myeloid<- expand.grid(Sender=as.list(tcell_types), Receiver=as.list(myeloid_types))
myeloid_tcell<- expand.grid(Sender=as.list(myeloid_types), Receiver=as.list(tcell_types))
tumor_tcell<- expand.grid(Sender=as.list(tumor_types), Receiver=as.list(tcell_types))
tumor_myeloid<- expand.grid(Sender=as.list(tumor_types), Receiver=as.list(myeloid_types))

## For tumor/tcell, split into clonal/nonclonal
tumor_clonal<- tumor_tcell[grepl("_clonal", tumor_tcell$Receiver),]
tumor_nonclonal<- tumor_tcell[grepl("_nonclonal", tumor_tcell$Receiver),]

## All together:
all_comparisons<- do.call("rbind", list(tcell_myeloid, myeloid_tcell, tumor_clonal, tumor_nonclonal, tumor_myeloid))
all_comparisons<- as.data.frame(apply(all_comparisons, 2, function(x){as.character(x)}))

  ## Add column with analysis name
  all_comparisons$name<- paste0(all_comparisons$Sender, "_to_", all_comparisons$Receiver)
  all_comparisons<- all_comparisons[,c("name", "Sender", "Receiver")]
  write.table(all_comparisons, file=paste0(analysis_dir, "all_comparisons.txt"), col.names = FALSE, quote = FALSE, row.names = FALSE)
  
## Separated to speed up run time (these files used)
comparison_list<- list(tcell_myeloid=tcell_myeloid, myeloid_tcell=myeloid_tcell, 
                       tumor_clonal=tumor_clonal, tumor_nonclonal=tumor_nonclonal, tumor_myeloid=tumor_myeloid)
comparison_list<- lapply(comparison_list, function(x){
  x$name<- paste0(x$Sender, "_to_", x$Receiver)
  x<- x[,c("name", "Sender", "Receiver")]
  x$Sender<- as.character(x$Sender);x$Receiver<- as.character(x$Receiver)
  return(x)
})
for(i in names(comparison_list)){
  tmp<- comparison_list[[i]]
  write.table(tmp, file=paste0(analysis_dir, "/byCellType/comparisons_", i, ".txt"), col.names = FALSE, quote = FALSE, row.names = FALSE)
}

```


## By program: tumor cell types, myeloid programs, cd4 programs, cd8 programs
```{r}
all_seurat<- readRDS(paste0(analysis_dir, "seurat_tumor.tcell.myeloid.Rds"))

## Add on CD4/CD8 distinction
all_seurat$detailed_annot<- ifelse(grepl("CD4", all_seurat$immunePrograms_tumorCellTypes), "CD4",
                                   ifelse(grepl("CD8", all_seurat$immunePrograms_tumorCellTypes), "CD8",
                                          ifelse(grepl("Unclear", all_seurat$immunePrograms_tumorCellTypes), "UnclearTcell",
                                                 all_seurat$highLevel_annot)))

## Remove "unclear" cells, inhibitory CD8
all_seurat<- subset(all_seurat, detailed_annot != "UnclearTcell")
all_seurat<- subset(all_seurat, immunePrograms_tumorCellTypes != "CD8_Inhibitory")

## Remove any programs/cell types without at least 10 cells in 1 sample
nCells<- as.data.frame.matrix(table(all_seurat$immunePrograms_tumorCellTypes, all_seurat$sample))
nSamples_pf<- apply(nCells, 1, function(x){sum(x>10)})
cellTypes_keep<- names(nSamples_pf)[nSamples_pf!=0]
all_seurat<- subset(all_seurat, immunePrograms_tumorCellTypes %in% cellTypes_keep)

## Get list of all programs by high level cell type
cd4_types<- unique(all_seurat$immunePrograms_tumorCellTypes[grepl("CD4", all_seurat$immunePrograms_tumorCellTypes)])
cd8_types<- unique(all_seurat$immunePrograms_tumorCellTypes[grepl("CD8", all_seurat$immunePrograms_tumorCellTypes)])
myeloid_types<- unique(all_seurat$immunePrograms_tumorCellTypes[all_seurat$detailed_annot=="Myeloid"])
tumor_types<- unique(all_seurat$immunePrograms_tumorCellTypes[all_seurat$highLevel_annot=="Tumor"])


## Between high level cell type lists, get all pairwise combinations
cd4_myeloid<- expand.grid(Sender=as.list(cd4_types), Receiver=as.list(myeloid_types))
cd8_myeloid<- expand.grid(Sender=as.list(cd8_types), Receiver=as.list(myeloid_types))
tumor_cd4<- expand.grid(Sender=as.list(tumor_types), Receiver=as.list(cd4_types))
tumor_cd8<- expand.grid(Sender=as.list(tumor_types), Receiver=as.list(cd8_types))
tumor_myeloid<- expand.grid(Sender=as.list(tumor_types), Receiver=as.list(myeloid_types))
all_list<- list(cd4_myeloid=cd4_myeloid, cd8_myeloid=cd8_myeloid, tumor_cd4=tumor_cd4, tumor_cd8=tumor_cd8, tumor_myeloid=tumor_myeloid)

## For each cell type list, cycle through each comparison. Remove any without at least 1 sample with 10 cells in each
nCells_allTypes<- as.data.frame.matrix(table(all_seurat$immunePrograms_tumorCellTypes, all_seurat$sample))
all_list_filtered<- lapply(all_list, function(x){
  ## Cycle through each pair in the comparisons table- check whether at least 1 sample "passes" (has 10 cells in each)
  pair_pf<- unlist(lapply(1:nrow(x), function(pair){
    tmp<- nCells_allTypes[unlist(x[pair,]),]
    pf<- sum(apply(tmp, 2, function(y){sum(y>10)==2})) ## number of samples with > 10 cells in BOTH cell types in pair
    if(pf!=0){return("pass")}else{return("fail")}
  }))
  
  ## Only return comparisons that "pass"
  return(x[pair_pf=="pass",])
})

## Add column with analysis name to each comparison table, then write out
all_list_filtered<- lapply(names(all_list_filtered), function(x){
  all_comparisons<- all_list_filtered[[x]]
  
  ## Convert columns to character- converted to list during pairwise creation
  all_comparisons$Sender<- as.character(all_comparisons$Sender)
  all_comparisons$Receiver<- as.character(all_comparisons$Receiver)
  
  ## Add on "comparison name"
  all_comparisons$name<- paste0(all_comparisons$Sender, "_to_", all_comparisons$Receiver)
  all_comparisons<- all_comparisons[,c("name", "Sender", "Receiver")]
  
  ## Write out to file for input into cellchat
  write.table(all_comparisons, file=paste0(analysis_dir, "byProgram/comparisons_byPrograms_",x, ".txt"), col.names = FALSE, quote = FALSE, row.names = FALSE)
})



```

## Generating comparisons file input, updated as of 01/18/22
Previously 1 batch submitted for each comparisons file on O2. To speed up computational time, updated to submit 1 batch for each pair
Generate updated comparisons file- all pairs in one file, no analysis name
```{r}
all_seurat<- readRDS(paste0(analysis_dir, "seurat_tumor.tcell.myeloid.Rds"))

## Vector of all programs
programs<- unique(all_seurat$immunePrograms_tumorCellTypes)
programs<- programs[!grepl("Unclear", programs)]

## All combinations
pairs<- apply(combn(programs,2),2,paste,collapse='__')
pairs<- data.frame(celltype1=sapply(pairs, function(x){unlist(strsplit(x, "__"))[1]}),
                   celltype2=sapply(pairs, function(x){unlist(strsplit(x, "__"))[2]}),
                   row.names = NULL)

## Get number of cells per program per sample
nCells<- as.data.frame(table(all_seurat$sample, all_seurat$immunePrograms_tumorCellTypes))
colnames(nCells)<- c("Sample", "Program", "NCells")
nCells<- nCells[nCells$NCells>10,c("Sample", "Program")]

## List of samples with programs in each
sample_list<- as.list(as.character(unique(nCells$Sample)))
sample_list<- lapply(sample_list, function(x){as.character(nCells$Program[nCells$Sample==x])})
names(sample_list)<- as.character(unique(nCells$Sample))

## Remove combinations without >10 cells in both celltypes in at least 1 sample
pairs_pf<- apply(pairs, 1, function(x){
  sum(unlist(lapply(sample_list, function(s){x["celltype1"] %in% s & x["celltype2"] %in% s})))
})
pairs_pf<- pairs[pairs_pf != 0,]

## Write out
write.table(pairs_pf, file=paste0(analysis_dir, "byProgram/all_comparisons_byPrograms.txt"), col.names = FALSE, quote = FALSE, row.names = FALSE)

```


```{r}
## How many cells are actually input into cellchat for each interaction? 
## To simplify analysis, remove an interactions that do not have at least 1 sample with >10 cells in both interactor
## These interactions will not be run in cell chat anyways
## Cycle through each sample, determine whether it has enough cells to run each interaction
cell_counts<- lapply(all_seurat_list, function(x){
  nCells<- as.data.frame(table(x$detailed_annotForCellChat))
  comp_counts<- lapply(comparisons$V1, function(comp){
    tmp<- comparisons[comparisons$V1==comp,]
    PF<- sum(nCells[nCells$Var1==tmp$V2, "Freq"]>10 & nCells[nCells$Var1==tmp$V3, "Freq"]>10)==1
    if(PF){
      total_counts<- nCells[nCells$Var1==tmp$V2, "Freq" ] + nCells[nCells$Var1==tmp$V3, "Freq" ] 
    }else{total_counts<-0}
    return(total_counts)
  })
  names(comp_counts)<- comparisons$V1
  comp_counts<- lapply(comp_counts, function(x){ifelse(length(x)==0, 0, x)})
  comp_counts<- as.data.frame(unlist(comp_counts))
  colnames(comp_counts)<-unique(x$sample)
  return(comp_counts)
})
cell_counts<- do.call("cbind", cell_counts)

## Number of samples that would be run for each interaction
nSamples_interaction<- apply(cell_counts,1, function(x){sum(x>0)})

## Interactions with 0 samples that would not be run- remove here for simplicity
interactions_remove<- names(nSamples_interaction)[nSamples_interaction==0]
comparisons<- comparisons[!comparisons$V1 %in% interactions_remove,]
write.table(all_comparisons, file=paste0(analysis_dir, "all_comparisons.txt"), col.names = FALSE, quote = FALSE, row.names = FALSE)
```
