==> ImmuneCellAnalysis/04a_Integrate_Immune.Tumor/CellChat/04b_CellChat_Visualizations_byProgram.Rmd <==
---
title: "Plot cellchat"
output: html_document
---

## Plot cell chat results between immune programs, tumor cell types

```{r}
library(CellChat)
library(NMF)
library(circlize)
library(patchwork)
library(Seurat) 
library(ggh4x)
library(ggalluvial)

options(stringsAsFactors = FALSE)

base_dir = "/Users/jlabelle/"
script_dir = paste0(base_dir, "Dropbox (Partners HealthCare)/FilbinLab/data_analysis/Shared/Scripts/")
working_dir<- paste0(base_dir, 
                     "Dropbox (Partners HealthCare)/FilbinLab/",
                     "data_analysis/Projects/git_repos/Pedi_Immune/ImmuneCellAnalysis/04a_Integrate_Immune.Tumor/CellChat/")
analysis_dir<- paste0(working_dir, "/analysis/byProgram/")
figure_dir<- paste0(working_dir, "/figures/byProgram/")

cellchatHelper<- paste0(script_dir, "CellChat_HelperFunctions.R")
source(cellchatHelper)

setwd(working_dir)

## Load in list of all cell types across all subtypes
nCell_bySample<- readRDS( "analysis/nCellTypes_bySample.Rds")


library(RColorBrewer)
myPalette <- colorRampPalette(rev(brewer.pal(11, "Spectral")))
```

## Settings
```{r}
## Remove BCOR?
remove_bcor<- FALSE

## Remove epn?
remove_epn<-FALSE

## Just epn (for supplemental)
just_epn<- TRUE

## Use downsampled results?
downsampled<-TRUE

## Set new figure dir
if(remove_bcor & remove_epn & !just_epn){
  if(downsampled){
    figure_dir_use<- paste0(figure_dir, "downsampled/removeBCOR.EPN/")
  }else{
    figure_dir_use<- paste0(figure_dir, "removeBCOR.EPN/")
  }
  
} else if(just_epn){
  if(downsampled){
    figure_dir_use<- paste0(figure_dir, "downsampled/onlyEPN/")
  }else{
    figure_dir_use<- paste0(figure_dir, "onlyEPN/")
  }
  
} else{
  if(downsampled){
    figure_dir_use<- paste0(figure_dir, "downsampled/")
  }else{
    figure_dir_use<- figure_dir
  }
}

print(paste0("FIGURE DIRECTORY IN USE: ", figure_dir_use))
if(!dir.exists(figure_dir_use)){dir.create(figure_dir_use, recursive = TRUE)}

## Set analysis dir
if(downsampled){analysis_dir_use<- paste0(analysis_dir, "downsampled/")}
if(!downsampled){analysis_dir_use<- paste0(analysis_dir, "all_cells/")}
```

## Load in LR results for all analyses (already merged)
```{r}
allLRs<- readRDS(paste0(analysis_dir_use, "all_LRs.Rds")) 

## Subset to just sig- by pvalue and prob
sigLRs<- allLRs[allLRs$pvalue<=0.1 & allLRs$Prob>0,]

## Add spacing to subtypes for better plotting
sigLRs$Subtype<- gsub("Histone", "\nHistone\n", gsub("Ependymoma", "EPN", sigLRs$Subtype))

## Remove bcor/epn if needed
if(remove_bcor){
  sigLRs<- sigLRs[sigLRs$Subtype != "hemiBCOR",]
}

if(remove_epn & !just_epn){
  sigLRs<- sigLRs[sigLRs$Subtype != "EPN",]
}

if(just_epn){
  sigLRs<- sigLRs[sigLRs$Subtype == "EPN",]
}

```



## 2 layers of cell type classification:
"major" cell type = CD4, CD8, Myeloid, or Tumor
"minor" cell type = Tumor cell type (AC-like, OPC-like), myeloid program, CD4 program, CD8 program
  

## Set minor cell types within major cell type, add onto sigLRs
```{r}
cd4_types<- as.character(unique(sigLRs$sender)[grepl("CD4", unique(sigLRs$sender))])
cd8_types<- as.character(unique(sigLRs$sender)[grepl("CD8", unique(sigLRs$sender))])
myeloid_types<- c("IFN_TAM", "Monocytes", "Inflammatory", "Hypoxic_TAM", "Stress_Response", "SEPP1_Mo_TAM")
tumor_types<- as.character(unique(sigLRs$sender)[!unique(sigLRs$sender) %in% c(cd4_types, cd8_types, myeloid_types)])
all_cell_types<- list(CD4=cd4_types, CD8=cd8_types, Myeloid=myeloid_types, Tumor=tumor_types)

sigLRs$sender_major<- ifelse(sigLRs$sender %in% tumor_types, "Tumor",
                      ifelse(sigLRs$sender %in% myeloid_types, "Myeloid",
                      ifelse(sigLRs$sender %in% cd4_types, "CD4", "CD8")))
sigLRs$receiver_major<- ifelse(sigLRs$receiver %in% tumor_types, "Tumor",
                        ifelse(sigLRs$receiver %in% myeloid_types, "Myeloid",
                        ifelse(sigLRs$receiver %in% cd4_types, "CD4", "CD8")))

## Add \n to minor cell types for better plotting
sigLRs$sender<- gsub("_", "\n", sigLRs$sender)
sigLRs$receiver<- gsub("_", "\n", sigLRs$receiver)
```


## General plot types:
1. all LRs between 2 major cell types, regardless of minor cell types
2. All LRs between 1 major cell type and minor cell types
3. All LRs between 2 major cell types' minor cell types- heatmap colored by % of samples that LR is found in
4. Number of interactions between all pairwise combinations of 2 sets of minor cell types
5. Summary plots for LRs of interest


## 1. all LRs between 2 major cell types, regardless of minor cell types
```{r}
celltypes<- c("Tumor", "Myeloid", "CD4", "CD8")
sr_pairs<-with(subset(expand.grid(celltypes,celltypes),Var1!=Var2),paste0(Var1,'-',Var2))

for(p in sr_pairs){
  ## Set major sender/receiver cell types
  majorSender<- unlist(strsplit(p, split="-"))[1]
  majorReceiver<- unlist(strsplit(p, split="-"))[2]
  sigLRs_plot<-sigLRs[sigLRs$sender_major==majorSender & sigLRs$receiver_major==majorReceiver,]
  
  ## Add column for unique ID- by sample/ligand/receptor
  sigLRs_plot$UniqueID<- paste0(sigLRs_plot$Sample, "_", sigLRs_plot$ligand, "_",sigLRs_plot$receptor)
  
  ## remove duplicate interactions
  sigLRs_unique<- merge(aggregate(Prob ~ UniqueID, sigLRs_plot, max), sigLRs_plot)
  sigLRs_unique<- sigLRs_unique[!duplicated(sigLRs_unique$UniqueID),] ## removes any total duplicates (identical prob too)
  
  ## Set variable to facet wrap by- usually will be subtype -  and plot
  sigLRs_unique$FacetWrap1<- sigLRs_unique$Subtype
  
  sigLRs_unique$LR<- sigLRs_unique$interaction_name
  PlotPvalueProb(sigLRs_unique,xAxis = "sample", labelXAxis = TRUE, facetWrap = c("Subtype"))+
    theme(axis.text = element_text(face="bold"))+
    ggtitle("")+
    theme(panel.grid.major.x = element_blank() ,panel.grid.major.y = element_line( size=.1, color="black" ) )+
    ggtitle(paste0("Sender: ", majorSender, "\nReceiver: ", majorReceiver))
  
  
  ## get appropriate width/height
  height<- 3+length(unique(sigLRs_unique$LR))*.1
  width<- 5+length(unique(sigLRs_unique$sample)) * .25
  
  ## save
  ggsave(paste0(figure_dir_use, "Sender.", majorSender, "_Receiver.", majorReceiver, ".png"), width=width, height=height)
}


```


## 2. All LRs between 1 major cell type and minor cell types
```{r}
celltypes<- c("Tumor", "Myeloid", "CD4", "CD8")
sr_pairs<-with(subset(expand.grid(celltypes,celltypes),Var1!=Var2),paste0(Var1,'-',Var2))

for(p in sr_pairs){
  ## Set major sender/receiver cell types
  majorSender<- unlist(strsplit(p, split="-"))[1]
  majorReceiver<- unlist(strsplit(p, split="-"))[2]
  
  ## Subset to sender/receiver major cell type
  sigLRs_plot<-sigLRs[sigLRs$sender_major==majorSender & sigLRs$receiver_major==majorReceiver,]
  
  ## Add column for unique ID- by sample/ligand/receptor/receiver cell type
  sigLRs_plot$UniqueID<- paste0(sigLRs_plot$Sample, "_", sigLRs_plot$ligand, "_",
                                sigLRs_plot$receptor, "_",sigLRs_plot$receiver)
  
  ## remove duplicate interactions
  sigLRs_unique<- merge(aggregate(Prob ~ UniqueID, sigLRs_plot, max), sigLRs_plot)
  sigLRs_unique<- sigLRs_unique[!duplicated(sigLRs_unique$UniqueID),] ## removes any total duplicates (identical prob too)
  
  ## Set variable to facet wrap by and plot
  sigLRs_unique$FacetWrap1<- sigLRs_unique$receiver
  
  sigLRs_unique$LR<- sigLRs_unique$interaction_name
  PlotPvalueProb(sigLRs_unique,xAxis = "sample", labelXAxis = TRUE, facetWrap = c( "Receiver"))+
    theme(axis.text = element_text(face="bold"))+
    ggtitle("")+
    theme(panel.grid.major.x = element_blank() ,panel.grid.major.y = element_line( size=.1, color="black" ) )+
    ggtitle(paste0("Sender: ", majorSender, "\nReceiver: ", majorReceiver))
  
  
  ## get appropriate width/height
  height<- 3+length(unique(sigLRs_unique$LR))*.1
  width<- 5+length(unique(sigLRs_unique$Sample)) * .25
  
  ## save
  ggsave(paste0(figure_dir_use, "Sender.", majorSender, "_Receiver.All",majorReceiver, ".png"), width=width, height=height)
}



```


## 3. All LRs between 2 major cell types' minor cell types- heatmap colored by % of samples that LR is found in
```{r}
celltypes<- c("Tumor", "Myeloid", "CD4", "CD8")
sr_pairs<-with(subset(expand.grid(celltypes,celltypes),Var1!=Var2),paste0(Var1,'-',Var2))

for(p in sr_pairs){
  ## Set major sender/receiver cell types
  majorSender<- unlist(strsplit(p, split="-"))[1]
  majorReceiver<- unlist(strsplit(p, split="-"))[2]
  
  ## Subset to sender/receiver major cell type
  sigLRs_plot<- sigLRs
  sigLRs_plot<-sigLRs_plot[sigLRs_plot$sender_major==majorSender & sigLRs_plot$receiver_major==majorReceiver,]
  
  ## Remove \n from sender cell types for better plotting
  sigLRs_plot$sender<- gsub("\n", "_", sigLRs_plot$sender)
  
  ## For each unique sender/receiver combination, determine the total number of samples tested
  sigLRs_plot$interaction_id<- paste0(sigLRs_plot$sender, "_", sigLRs_plot$receiver)
  sigLRs_plot<- sigLRs_plot %>% group_by(interaction_id) %>% 
    mutate(nSamples_inInteraction=length(unique(sample))) %>% as.data.frame()
  
  ## For each unique LR/interaction, determine the total number of samples identified
  sigLRs_plot$UniqueID<- paste0(sigLRs_plot$ligand, "_", sigLRs_plot$receptor, "_", sigLRs_plot$interaction_id)
  sigLRs_plot<- sigLRs_plot %>% group_by(UniqueID) %>% mutate(nSamples_inLR=length(unique(sample))) %>% as.data.frame()
  
  ## For each unique LR/interaction, divide n samples with LR/ n samples in interacting pair
  sigLRs_unique<- sigLRs_plot[!duplicated(sigLRs_plot$UniqueID),]
  sigLRs_unique$perSamples_inLR<- sigLRs_unique$nSamples_inLR/sigLRs_unique$nSamples_inInteraction
  
  ## Add on ID for ligand/receptor
  sigLRs_unique$LR<- paste0(sigLRs_unique$ligand, "_", sigLRs_unique$receptor)
  
  ggplot(sigLRs_unique, aes(x=sender, y=LR, fill=perSamples_inLR))+
    geom_tile()+
    theme_linedraw() + 
    facet_grid(cols = vars(receiver), scales = "free_x", space="free_x")+
    scale_fill_gradient(low="white", high="red")+
    theme(axis.text.x = element_text(angle=45, hjust=1),
          axis.text=element_text(color="black", face="bold"),
          axis.title = element_text(color="black", face="bold"))+
    ylab("") + xlab("Sender cell type")
  
  height<- 3+length(unique(sigLRs_unique$LR))*.1
  width<- 5+length(unique(sigLRs_unique$sender)) * .25
  ggsave(paste0(figure_dir_use, "HeatmapBySampleProportion_Sender.", majorSender, "_Receiver.", majorReceiver, ".png"),
         width=width, height=height)
  
}


```

## 4. Number of interactions between all pairwise combinations of 2 sets of minor cell types
Plotting options: sankey plot (USED DOWNSTREAM), heatmap, chord diagram

## Preprocess
```{r}
## subset to unique LRs in pair
sigLRs$UniqueID<- paste0(sigLRs$sender, ".to.", sigLRs$receiver, "_", sigLRs$interaction_name)
sigLRs<- sigLRs[!duplicated(sigLRs$UniqueID),]

## Get number of interactions per pair
bd<- as.data.frame(table(sigLRs$sender, sigLRs$receiver))
colnames(bd)<- c("sender", "receiver", "nLR")


## Add major sender/receiver
bd$major_sender<- as.character(plyr::mapvalues(bd$sender, sigLRs$sender, sigLRs$sender_major, warn_missing = FALSE))
bd$major_receiver<- as.character(plyr::mapvalues(bd$receiver, sigLRs$receiver, sigLRs$receiver_major, warn_missing = FALSE))



## Order cell types by major cell type
ordered_celltypes<- gsub("\n", "_", c(cd4_types, cd8_types, myeloid_types, tumor_types))
sigLRs_plot$receiver.plot<- factor(sigLRs_plot$receiver, levels=ordered_celltypes[ordered_celltypes %in% sigLRs_plot$receiver])
sigLRs_plot$sender.plot<- factor(sigLRs_plot$sender, levels=ordered_celltypes[ordered_celltypes %in% sigLRs_plot$sender])
#sigLRs_plot$major_sender<- gsub("CD4", "Immune", gsub("CD8", "Immune",gsub("Myeloid", "Immune", sigLRs_plot$major_sender)))
```


## Sankey plot
```{r}
sigLRs_plot <-bd

## Plot with my helper function
sr_pairs<- list(list(c("CD4", "CD8"), c("Tumor")),
                list(c("CD4", "CD8"), c("Myeloid")),
                list(c("Myeloid"), c("Tumor")),
                list(c("CD4"), c("CD8")))
sigLRs_plot$nLRs_norm<- sigLRs_plot$nLR

all_plots<- lapply(sr_pairs, function(p){
  tmp<- sigLRs_plot[(sigLRs_plot$major_receiver %in% p[[1]] & sigLRs_plot$major_sender %in% p[[2]]) |
                    (sigLRs_plot$major_receiver %in% p[[2]] & sigLRs_plot$major_sender %in% p[[1]]), ]
  if(length(p[[1]])>1 | length(p[[2]]) > 1){ ## simplify to combine cd4/cd8 if needed
    tmp$major_receiver<- gsub("CD4", "Immune", gsub("CD8", "Immune", tmp$major_receiver))
    tmp$major_sender<- gsub("CD4", "Immune", gsub("CD8", "Immune", tmp$major_sender))
  }
  
  my_sankeyPlot(tmp)
})
cowplot::plot_grid(plotlist=all_plots, ncol=length(sr_pairs))
ggsave(paste0(figure_dir_use, "Sankey_LRs_withCD4.CD8.pdf"), width=length(sr_pairs)*10, height=25)

```




## Heatmap
```{r}
sigLRs_plot<- bd

## Switch sender/receiver for one- just for plotting purposes
sigLRs_plot$sender.plot<- ifelse(sigLRs_plot$major_receiver==majorCellType1, sigLRs_plot$receiver, sigLRs_plot$sender)
sigLRs_plot$receiver.plot<- ifelse(sigLRs_plot$major_receiver==majorCellType1, sigLRs_plot$sender, sigLRs_plot$receiver)

## Order by major cell type
sigLRs_plot$sender.plot<- factor(sigLRs_plot$sender.plot, levels=c(majorCells1, majorCells2))
sigLRs_plot$receiver.plot<- factor(sigLRs_plot$receiver.plot, levels=c(majorCells1, majorCells2))

## Plot
ggplot(sigLRs_plot, aes(x=sender.plot, y=receiver.plot, fill=nLRs_norm))+
  geom_tile()+
  #theme_linedraw() + 
  theme_classic()+
  scale_fill_gradient(low="white", high="red")+
  theme(axis.text.x = element_text(angle=45, hjust=1),
        axis.text=element_text(color="black", face="bold"),
        axis.title = element_text(color="black", face="bold"),
        strip.text = element_text(face="bold", size=12))+
  ylab("") + xlab("")+
  facet_grid(cols=vars(major_sr))

ggsave(paste0(figure_dir_use, "HeatmapByTotalLR_Norm", norm_method,"_",paste(majorCellType1, collapse="."),".", majorCellType2, ".png"),
       width=6, height=4)

```


## chord diagram
```{r}
sigLRs_plot<- sigLRs_nLR

## Alternative: chord diagram
majorCells1<- gsub("_", "\n",gsub("\\.", "\n",majorCells1))
majorCells2<- gsub("_", "\n",gsub("\\.", "\n",majorCells2))

  ## Set colors- shades of blue and green
  colfunc1 <- colorRampPalette(c("olivedrab", "cyan"))
  colfunc2<- colorRampPalette(c("orange", "violetred"))
  colors_use<-c(colfunc1(length(majorCells1)),colfunc2(length(majorCells2)))
  names(colors_use)<- c(majorCells1, majorCells2)
  
  ## Subset to needed columns
  tmp<- sigLRs_plot[,c( "sender", "receiver","nLRs_norm")]
  tmp$sender<- gsub( "_","\n",gsub("\\.", "\n", tmp$sender))
  tmp$receiver<- gsub("_", "\n", gsub("\\.", "\n",tmp$receiver))
  
  ## Plot
  png(paste0(figure_dir_use, "ChordByTotalLR_Norm", norm_method,"_",paste(majorCellType1, collapse="."),".", majorCellType2, ".png"), width=400, height=400) ## for png, use size of 1000
  chordDiagram(tmp,annotationTrack = c("grid", "name"), order=c(majorCells1, majorCells2),
                       grid.col = colors_use)
  dev.off()
```






## 5. Summary plots for LRs of interest
```{r}
## Set major sender/receiver cell types +  ligand/receptors of interest
majorSender<- "Tumor"
majorReceiver<- "Myeloid" ## Tcell, CD4, CD8, or Myeloid
LR_list<- list(Tumor.Tcell.CD4.CD8=c("PVR_TIGIT","PROS1_AXL","CXCL16_CXCR6","CDH1_KLRG1","ANGPTL2_LILRB3", "CLEC2D_KLRB1"),
               Tumor.Myeloid=c("CCL5_CCR1","PROS1_AXL","ANGPTL2_LILRB3"))
LRs<- unlist(unname(LR_list[grepl(majorReceiver, names(LR_list))]))

## If using Tcell, remove clonality info
sigLRs_plot<- sigLRs
sigLRs_plot$sender<- gsub("_nonclonal", "", gsub("_clonal", "", sigLRs_plot$sender))
sigLRs_plot$receiver<- gsub("_nonclonal", "", gsub("_clonal", "", sigLRs_plot$receiver))


## Subset to all Tcells, CD4, or CD8
if(majorReceiver=="Tcell"){
  sigLRs_plot<-sigLRs_plot[sigLRs_plot$sender_major==majorSender & sigLRs_plot$receiver_major==majorReceiver,]
}else{
  sigLRs_plot<-sigLRs_plot[sigLRs_plot$sender_major==majorSender & sigLRs_plot$receiver==majorReceiver,]
}

## Subset to LR pairs of interest
sigLRs_plot$LR<- paste0(sigLRs_plot$ligand, "_", sigLRs_plot$receptor)
sigLRs_plot<- sigLRs_plot[sigLRs_plot$LR %in% LRs,]

## If using Tcell, remove clonality info
sigLRs_plot$sender<- gsub("_nonclonal", "", gsub("_clonal", "", sigLRs_plot$sender))
sigLRs_plot$receiver<- gsub("_nonclonal", "", gsub("_clonal", "", sigLRs_plot$receiver))

## Add column for unique ID- by sample/ligand/receptor
sigLRs_plot$UniqueID<- paste0(sigLRs_plot$Sample, "_", sigLRs_plot$ligand, "_",sigLRs_plot$receptor)

## remove duplicate interactions
sigLRs_unique<- merge(aggregate(Prob ~ UniqueID, sigLRs_plot, max), sigLRs_plot)
sigLRs_unique<- sigLRs_unique[!duplicated(sigLRs_unique$UniqueID),] ## removes any total duplicates (identical prob too)

## Set variable to facet wrap by- usually will be subtype -  and plot
sigLRs_unique$FacetWrap1<- sigLRs_unique$Subtype

PlotPvalueProb(sigLRs_unique,xAxis = "Sample", labelXAxis = TRUE, facetWrap = c("Subtype"))+
  theme(axis.text = element_text(face="bold"))+
  ggtitle("")+
  theme(panel.grid.major.x = element_blank() ,panel.grid.major.y = element_line( size=.1, color="black" ) )+
  ggtitle(paste0("Sender: ", majorSender, "\nReceiver: ", majorReceiver))


## get appropriate width/height
height<- 3+length(unique(sigLRs_unique$LR))*.1
width<- 5+length(unique(sigLRs_unique$Sample)) * .25

## save
ggsave(paste0(figure_dir, "LR.summary_Sender.", majorSender, "_Receiver.", majorReceiver, ".png"), width=width, height=height)
```
